+++ examples/client/files_examples.py
@@ -1,0 +1,5 @@
+import nc_py_api
+
+
+if __name__ == "__main__":
+    nc_py_api.Nextcloud()