+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.0.26 - 2023-0x-xx]
+## [0.0.26 - 2023-07-xx]
 
 ### Added
 
-- First Examples.
 - More documentation.
 - First `Notifications` APIs.
+
+### Changed
+
+- Reworked `User Status API`
+- Reworked return type for `weather_status.get_location`
 
 ## [0.0.25 - 2023-07-25]